SAV-5384 address comments: move to layout folder
