+++ src/library/index.ts
@@ -21,7 +21,6 @@ import RcSesButton from '@/components/common/Button'
 import RcSesButtonWithPopover from '@/components/common/Button/ButtonWithPopover'
 import RcSesCard from '@/components/common/Card'
 import DataPagination from '@/components/common/DataPagination'
-import RcSesFooter from '@/components/common/Footer'
 import RcSesIconWithCircularBackground from '@/components/common/IconWithCircularBackground'
 import RcSesTab from '@/components/common/Tabs/Tab'
 import RcSesTabPanel from '@/components/common/Tabs/TabPanel'
@@ -40,6 +39,7 @@ import RcSesRadioButtonGroup from '@/components/form/inputs/RadioButtonGroup'
 import RcSesSearchableField from '@/components/form/inputs/SearchableField'
 import RcSesSelect from '@/components/form/inputs/Select'
 import RcSesTextField from '@/components/form/inputs/TextField'
+import RcSesFooter from '@/components/layout/Footer'
 import RcSesServiceFormActions from '@/components/layout/ServiceFormActions'
 import RcSesServiceFormContainer from '@/components/layout/ServiceFormContainer'
 import RcSesServiceHeader from '@/components/layout/ServiceHeader'
